requirements.txt
gunicorn==20.1.0
Django==4.2.9
django-allauth==0.58.2
django-bootstrap5==23.4
django-cleanup==8.1.0
django-countries==7.6.1
django-crispy-forms==2.3
django-filter==23.3
django-jazzmin==3.0.0
django-taggit==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
stripe==10.3.0
redis==5.0.1
celery==5.3.6
pdfkit==1.0.0
weasyprint==62.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.2
psycopg2-binary==2.9.9
fasttext-wheel==0.9.2
gensim==4.3.2
spacy==3.7.2
transformers==4.41.2
python-decouple==3.8
python-dotenv==1.0.1
whitenoise==6.7.0
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.0
xhtml2pdf==0.2.16
groq==0.6.0
cloudscraper==1.2.71
